backend/requirements.txt
numpy
fastapi
uvicorn[standard]
langchain
langchain-openai
langchain-anthropic
langchain-huggingface
langchain-community
langchain-chroma
langchain-text-splitters
chromadb
sentence-transformers
transformers
faiss-cpu
pypdf
pypandoc
python-multipart
requests
python-dotenv
tiktoken
